complete homework 3

+++ src/main/java/ru/extoozy/service/profile/impl/UserProfileServiceImpl.java
@@ -5,7 +5,9 @@ import ru.extoozy.context.UserContext;
 import ru.extoozy.dto.profile.CreateUserProfileDto;
 import ru.extoozy.dto.profile.UpdateUserProfileDto;
 import ru.extoozy.dto.profile.UserProfileDto;
+import ru.extoozy.entity.UserEntity;
 import ru.extoozy.entity.UserProfileEntity;
+import ru.extoozy.enums.UserRole;
 import ru.extoozy.exception.ResourceNotFoundException;
 import ru.extoozy.mapper.UserProfileMapper;
 import ru.extoozy.repository.profile.UserProfileRepository;
@@ -20,7 +22,7 @@ public class UserProfileServiceImpl implements UserProfileService {
 
     @Override
     public void create(CreateUserProfileDto dto) {
-        UserProfileEntity userProfile = UserProfileMapper.toEntity(dto);
+        UserProfileEntity userProfile = UserProfileMapper.INSTANCE.toEntity(dto);
         userProfile.setUser(UserContext.getUser());
         userProfileRepository.save(userProfile);
         UserContext.getUser().setUserProfile(userProfile);
@@ -28,34 +30,73 @@ public class UserProfileServiceImpl implements UserProfileService {
 
     @Override
     public void update(UpdateUserProfileDto dto) {
-        UserProfileEntity userProfile = UserProfileMapper.toEntity(dto);
+        UserProfileEntity userProfile = UserProfileMapper.INSTANCE.toEntity(dto);
+        if (usesDoesNotHaveAccessToUserProfile(UserContext.getUser(), userProfileRepository.findById(dto.getId()))) {
+            throw new ResourceNotFoundException(
+                    "The user with id=%s does not have access to user profile with id=%s".formatted(
+                            UserContext.getUser().getId(),
+                            userProfile.getId()
+                    )
+            );
+        }
+
         userProfileRepository.update(userProfile);
     }
 
     @Override
     public UserProfileDto get(Long id) {
         UserProfileEntity userProfile = userProfileRepository.findById(id);
-        return UserProfileMapper.toDto(userProfile);
+        if (usesDoesNotHaveAccessToUserProfile(UserContext.getUser(), userProfile)) {
+            throw new ResourceNotFoundException(
+                    "The user with id=%s does not have access to user profile with id=%s".formatted(
+                            UserContext.getUser().getId(),
+                            userProfile.getId()
+                    )
+            );
+        }
+        return UserProfileMapper.INSTANCE.toDto(userProfile);
     }
 
     @Override
     public UserProfileDto getByUserId(Long userId) {
         UserProfileEntity userProfile = userProfileRepository.findByUserId(userId);
-        return UserProfileMapper.toDto(userProfile);
+        if (usesDoesNotHaveAccessToUserProfile(UserContext.getUser(), userProfile)) {
+            throw new ResourceNotFoundException(
+                    "The user with id=%s does not have access to user profile with id=%s".formatted(
+                            UserContext.getUser().getId(),
+                            userProfile.getId()
+                    )
+            );
+        }
+        return UserProfileMapper.INSTANCE.toDto(userProfile);
     }
 
 
     @Override
     public List<UserProfileDto> getAll() {
         List<UserProfileEntity> userProfiles = userProfileRepository.findAll();
-        return UserProfileMapper.toDto(userProfiles);
+        return UserProfileMapper.INSTANCE.toDto(userProfiles);
     }
 
     @Override
     public void delete(Long id) {
+        UserProfileEntity userProfile = userProfileRepository.findById(id);
+        if (usesDoesNotHaveAccessToUserProfile(UserContext.getUser(), userProfile)) {
+            throw new ResourceNotFoundException(
+                    "The user with id=%s does not have access to user profile with id=%s".formatted(
+                            UserContext.getUser().getId(),
+                            userProfile.getId()
+                    )
+            );
+        }
         boolean deleted = userProfileRepository.delete(id);
         if (!deleted) {
             throw new ResourceNotFoundException("User profile with id=%s not found".formatted(id));
         }
+    }
+
+    private boolean usesDoesNotHaveAccessToUserProfile(UserEntity user, UserProfileEntity userProfile) {
+        return !user.getRole().equals(UserRole.ADMIN) &&
+                !user.getUserProfile().getId().equals(userProfile.getId());
     }
 }
