+++ src/main/java/ru/extoozy/service/user/impl/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.extoozy.service.user.impl;
 
 import lombok.RequiredArgsConstructor;
+import ru.extoozy.context.UserContext;
 import ru.extoozy.dto.user.AuthUserDto;
 import ru.extoozy.dto.user.UpdateUserDto;
 import ru.extoozy.dto.user.UserDto;
@@ -14,6 +15,7 @@ import ru.extoozy.security.PasswordHelper;
 import ru.extoozy.service.user.UserService;
 import ru.extoozy.util.RegexUtil;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RequiredArgsConstructor
@@ -21,39 +23,68 @@ public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
+
     @Override
     public void create(AuthUserDto dto) {
-        UserEntity user = UserMapper.toEntity(dto);
+        UserEntity user = UserMapper.INSTANCE.toEntity(dto);
         user.setRole(UserRole.USER);
         userRepository.save(user);
     }
 
     @Override
     public void update(UpdateUserDto dto) {
+        UserEntity user = UserMapper.INSTANCE.toEntity(dto);
+        if (userDoesNotHaveAccessToUser(UserContext.getUser(), user)) {
+            throw new ResourceNotFoundException(
+                    "The user with id=%s does not have access to user with id=%s".formatted(
+                            UserContext.getUser().getId(),
+                            user.getId()
+                    )
+            );
+        }
         if (RegexUtil.isInvalidEmail(dto.getEmail())) {
             throw new InvalidEmailException("Email=%s is invalid".formatted(dto.getEmail()));
         }
-        UserEntity entity = UserMapper.toEntity(dto);
         if (dto.getPassword() != null) {
-            entity.setPassword(PasswordHelper.getPasswordHash(dto.getPassword()));
+            user.setPassword(PasswordHelper.getPasswordHash(dto.getPassword()));
         }
-        userRepository.update(entity);
+        userRepository.update(user);
     }
 
     @Override
     public UserDto get(Long id) {
         UserEntity user = userRepository.findById(id);
-        return UserMapper.toDto(user);
+        if (userDoesNotHaveAccessToUser(UserContext.getUser(), user)) {
+            throw new ResourceNotFoundException(
+                    "The user with id=%s does not have access to user with id=%s".formatted(
+                            UserContext.getUser().getId(),
+                            user.getId()
+                    )
+            );
+        }
+        return UserMapper.INSTANCE.toDto(user);
     }
 
     @Override
     public List<UserDto> getAll() {
-        List<UserEntity> users = userRepository.findAll();
-        return UserMapper.toDto(users);
+        if (UserContext.getUser().getRole().equals(UserRole.ADMIN)) {
+            List<UserEntity> users = userRepository.findAll();
+            return UserMapper.INSTANCE.toDto(users);
+        }
+        return new ArrayList<>();
     }
 
     @Override
     public void delete(Long id) {
+        UserEntity user = userRepository.findById(id);
+        if (userDoesNotHaveAccessToUser(UserContext.getUser(), user)) {
+            throw new ResourceNotFoundException(
+                    "The user with id=%s does not have access to user with id=%s".formatted(
+                            UserContext.getUser().getId(),
+                            user.getId()
+                    )
+            );
+        }
         boolean deleted = userRepository.delete(id);
         if (!deleted) {
             throw new ResourceNotFoundException("User with id=%s not found".formatted(id));
@@ -72,6 +103,12 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void changeBlockStatus(Long id) {
-        userRepository.changeBlockStatus(id);
+        if (UserContext.getUser().getRole().equals(UserRole.ADMIN)) {
+            userRepository.changeBlockStatus(id);
+        }
+    }
+
+    private boolean userDoesNotHaveAccessToUser(UserEntity user, UserEntity accessingUser) {
+        return !user.getRole().equals(UserRole.ADMIN) && !user.getId().equals(accessingUser.getId());
     }
 }
